refactor(indexer): extract lookup methods into indexer/lookup.rs
Move 10 impl Indexer lookup methods and 2 free functions out of indexer.rs
into a new src/indexer/lookup.rs submodule, following the apply.rs pattern.
Move 5 related tests to src/indexer/lookup_tests.rs.

Extracted:
- is_declared_in, find_definition, find_definition_qualified
- hover_info, hover_info_at_location
- file_symbols, package_of, declared_package_of
- declared_parent_class_of, resolve_symbol_via_import
- symbol_kw, symbol_kw_for_lang (free functions)

All 453 tests pass.
