+++ src/indexer.rs
@@ -63,6 +63,8 @@ mod apply;
 pub(crate) use self::apply::{file_contributions, stale_keys_for, build_bare_names};
 use self::apply::hash_str;
 
+mod lookup;
+
 // Re-export cache/scan items needed by the inline test module below.
 #[cfg(test)]
 use self::cache::{FileCacheEntry, cache_entry_to_file_result};
@@ -309,14 +311,6 @@ impl Indexer {
         }
         None
     }
-
-    /// Returns true if `name` has at least one definition location inside `uri`.
-    pub fn is_declared_in(&self, uri: &Url, name: &str) -> bool {
-        self.definitions.get(name)
-            .map(|locs| locs.iter().any(|l| l.uri == *uri))
-            .unwrap_or(false)
-    }
-
     /// Like `word_at` but also returns the single dot-qualifier immediately
     /// preceding the word, if any.
     ///
@@ -395,24 +389,7 @@ impl Indexer {
         };
 
         Some((word, qualifier))
-    }
-
-    /// Resolve definition locations for `name` (with optional dot-qualifier).
-    #[allow(dead_code)]
-    pub fn find_definition(&self, name: &str, from_uri: &Url) -> Vec<Location> {
-        crate::resolver::resolve_symbol(self, name, None, from_uri)
-    }
-
-    pub fn find_definition_qualified(
-        &self,
-        name: &str,
-        qualifier: Option<&str>,
-        from_uri: &Url,
-    ) -> Vec<Location> {
-        crate::resolver::resolve_symbol(self, name, qualifier, from_uri)
-    }
-
-    /// If `name` at `position` is `it` or a named lambda parameter, return the
+    }    /// If `name` at `position` is `it` or a named lambda parameter, return the
     /// inferred element/receiver type name (e.g. `"Product"`, `"User"`).
     ///
     /// Used by hover and go-to-definition to provide useful info for lambda params.
@@ -788,66 +765,7 @@ impl Indexer {
         (items, hit_cap)
     }
 
-    /// Build a Markdown hover snippet for a symbol name.
-    pub fn hover_info(&self, name: &str) -> Option<String> {
-        // Check stdlib first so well-known symbols (run, apply, map, …) get
-        // proper signatures even when no project source contains them.
-        if let Some(md) = crate::stdlib::hover(name) { return Some(md); }
-
-        // Drop the dashmap ref before taking the second one.
-        let loc: Location = {
-            let r = self.definitions.get(name)?;
-            r.first()?.clone()
-        };
-        self.hover_info_at_location(&loc, name)
-    }
-
     /// Build hover markdown for `name` at a specific resolved `Location`.
-    /// Used by the hover handler so it shows the same symbol as go-to-definition.
-    pub fn hover_info_at_location(&self, loc: &Location, name: &str) -> Option<String> {
-        // On-demand index: the file may have been found by rg but not yet indexed.
-        if !self.files.contains_key(loc.uri.as_str()) {
-            if let Ok(path) = loc.uri.to_file_path() {
-                if let Ok(content) = std::fs::read_to_string(&path) {
-                    self.index_content(&loc.uri, &content);
-                }
-            }
-        }
-        let data = self.files.get(loc.uri.as_str())?;
-        // Prefer exact match by resolved location range; fall back to name match
-        // for symbols found via rg where the range may not align exactly.
-        let sym = data.symbols.iter().find(|s| s.selection_range == loc.range)
-            .or_else(|| data.symbols.iter().find(|s| s.name == name))?;
-
-        let start_line = sym.selection_range.start.line as usize;
-        let sig = collect_signature(&data.lines, start_line);
-
-        let lang = if loc.uri.path().ends_with(".kt") { "kotlin" }
-                   else if loc.uri.path().ends_with(".swift") { "swift" }
-                   else { "java" };
-
-        let code_block = if sig.is_empty() {
-            format!("```{}\n{} {}\n```", lang, symbol_kw_for_lang(sym.kind, lang), name)
-        } else {
-            format!("```{}\n{}\n```", lang, sig)
-        };
-
-        // Prepend KDoc / Javadoc comment if one immediately precedes the declaration.
-        if let Some(doc) = extract_doc_comment(&data.lines, start_line) {
-            Some(format!("{doc}\n\n---\n\n{code_block}"))
-        } else {
-            Some(code_block)
-        }
-    }
-
-    /// All symbols declared in the given file (for `documentSymbol`).
-    pub fn file_symbols(&self, uri: &Url) -> Vec<SymbolEntry> {
-        self.files
-            .get(uri.as_str())
-            .map(|d| d.symbols.clone())
-            .unwrap_or_default()
-    }
-
     /// Find the name of the innermost enclosing class/interface/object
     /// that contains `row` in the given file.
     ///
@@ -904,82 +822,6 @@ impl Indexer {
             }
         }
         None
-    }
-
-    /// Return the package declared in the given file, if any.
-    pub fn package_of(&self, uri: &Url) -> Option<String> {
-        self.files.get(uri.as_str())?.package.clone()
-    }
-
-    /// Return the package in which `name` is declared, by looking up its
-    /// definition locations and reading the `package` field of those files.
-    /// If `prefer_uri` is set, prefer definitions from that file first.
-    pub fn declared_package_of(&self, name: &str) -> Option<String> {
-        let locs = self.definitions.get(name)?;
-        for loc in locs.iter() {
-            if let Some(f) = self.files.get(loc.uri.as_str()) {
-                if let Some(pkg) = &f.package {
-                    return Some(pkg.clone());
-                }
-            }
-        }
-        None
-    }
-
-    /// If `name` is declared as an inner/nested class, return the name of its
-    /// enclosing class at the declaration site in `preferred_uri` (if found there),
-    /// otherwise the first definition site.
-    pub fn declared_parent_class_of(&self, name: &str, preferred_uri: &Url) -> Option<String> {
-        let locs = self.definitions.get(name)?;
-        // Try declaration in the preferred (current) file first.
-        for loc in locs.iter() {
-            if loc.uri == *preferred_uri {
-                return self.enclosing_class_at(&loc.uri, loc.range.start.line);
-            }
-        }
-        // Fall back to first definition in any file.
-        for loc in locs.iter() {
-            if let Some(parent) = self.enclosing_class_at(&loc.uri, loc.range.start.line) {
-                return Some(parent);
-            }
-        }
-        None
-    }
-
-    /// Scan imports in `uri` for `name` and return (parent_class, declared_pkg)
-    /// as resolved from the import statement.  E.g.:
-    ///   `import com.example.DashboardViewModel.Effect`
-    ///   → parent_class = Some("DashboardViewModel"), pkg = Some("com.example.DashboardViewModel")
-    pub fn resolve_symbol_via_import(
-        &self,
-        uri: &Url,
-        name: &str,
-    ) -> (Option<String>, Option<String>) {
-        let file = match self.files.get(uri.as_str()) {
-            Some(f) => f,
-            None    => return (None, None),
-        };
-        for line in file.lines.iter() {
-            let t = line.trim();
-            if !t.starts_with("import ") { continue; }
-            // Handle `import a.b.c.Name` and `import a.b.c.Name as Alias`
-            let import_path = t["import ".len()..].split_whitespace().next().unwrap_or("");
-            let segments: Vec<&str> = import_path.split('.').collect();
-            // Last segment should match `name` (or be `*`).
-            let last = *segments.last().unwrap_or(&"");
-            if last != name && last != "*" { continue; }
-
-            // Found a matching import. The declared package is everything up to (not incl.) `name`.
-            // The parent class is the segment immediately before `name` if it starts uppercase.
-            if last == name && segments.len() >= 2 {
-                let pkg = segments[..segments.len() - 1].join(".");
-                let parent = segments.get(segments.len() - 2)
-                    .filter(|s| s.chars().next().map(|c| c.is_uppercase()).unwrap_or(false))
-                    .map(|s| s.to_string());
-                return (parent, Some(pkg));
-            }
-        }
-        (None, None)
     }
 }
 
@@ -1101,28 +943,6 @@ fn callee_to_qualifier(full_callee: &str) -> Option<String> {
         .map(|s| s.to_string())
 }
 
-fn symbol_kw(kind: SymbolKind) -> &'static str {
-    match kind {
-        SymbolKind::CLASS          => "class",
-        SymbolKind::INTERFACE      => "interface",
-        SymbolKind::FUNCTION       => "fun",
-        SymbolKind::METHOD         => "fun",
-        SymbolKind::VARIABLE       => "var",
-        SymbolKind::CONSTANT       => "val",
-        SymbolKind::OBJECT         => "object",
-        SymbolKind::TYPE_PARAMETER => "typealias",
-        SymbolKind::ENUM           => "enum class",
-        SymbolKind::FIELD          => "field",
-        _                          => "symbol",
-    }
-}
-
-fn symbol_kw_for_lang(kind: SymbolKind, lang: &str) -> &'static str {
-    let kw = symbol_kw(kind);
-    // Swift uses `func`, not `fun`.
-    if lang == "swift" && kw == "fun" { "func" } else { kw }
-}
-
 // ─── tests ───────────────────────────────────────────────────────────────────
 
 #[cfg(test)]
@@ -1138,41 +958,6 @@ mod tests {
         let idx = Indexer::new();
         idx.index_content(&u, src);
         (u, idx)
-    }
-
-    // ── Swift hover uses "func" not "fun" ────────────────────────────────────
-
-    #[test]
-    fn swift_hover_uses_func_keyword() {
-        // Swift function with no signature detail should show "func", not "fun".
-        let src = "func greet() {}";
-        let (u, idx) = indexed("/Greeting.swift", src);
-        let hover = idx.hover_info_at_location(
-            &Location { uri: u.clone(), range: Default::default() },
-            "greet",
-        ).unwrap_or_default();
-        assert!(
-            hover.contains("func"),
-            "Swift hover should say 'func', got: {hover}"
-        );
-        assert!(
-            !hover.contains("```kotlin\nfun ") && !hover.contains("```swift\nfun "),
-            "Swift hover must not emit 'fun', got: {hover}"
-        );
-    }
-
-    #[test]
-    fn kotlin_hover_still_uses_fun_keyword() {
-        let src = "fun greet() {}";
-        let (u, idx) = indexed("/Greeting.kt", src);
-        let hover = idx.hover_info_at_location(
-            &Location { uri: u.clone(), range: Default::default() },
-            "greet",
-        ).unwrap_or_default();
-        assert!(
-            hover.contains("fun"),
-            "Kotlin hover should say 'fun', got: {hover}"
-        );
     }
 
     // ── word_at ──────────────────────────────────────────────────────────────
@@ -1815,21 +1600,6 @@ class Foo @Inject constructor(
         let sig = collect_signature(&lines, 0);
         assert!(!sig.contains('{'), "brace should be stripped");
         assert!(sig.contains("Foo"), "class name must be present");
-    }
-
-    #[test]
-    fn hover_includes_kdoc() {
-        let src = r#"package com.example
-
-/**
- * Represents a user account.
- */
-class Account(val name: String)"#;
-        let (u, idx) = indexed("/Account.kt", src);
-        let hover = idx.hover_info("Account").unwrap();
-        assert!(hover.contains("Represents a user account"), "got: {hover}");
-        assert!(hover.contains("```kotlin"), "got: {hover}");
-        assert!(hover.contains("---"), "separator missing: {hover}");
     }
 
     #[test]
